requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0
numpy>=1.19
scipy
scikit-image
opencv-python
imagehash
tensorboard
tqdm
pycocotools
